+++ py-xbbg/src/xbbg/_exports.py
@@ -9,6 +9,7 @@ SDK_EXPORTS = (
 CORE_EXPORTS = (
     "set_log_level",
     "get_log_level",
+    "enable_sdk_logging",
 )
 
 SERVICE_EXPORTS = (